Move loadSnippets() to main.ts
